requirements.txt
beautifulsoup4 == 4.7.1
celery == 4.3.0
django - extra - fields == 1.2.0
django - filter == 2.1.0
django - markdownify == 0.8.0
django - multiselectfield == 0.1.8
django - notifications - hq == 1.5.0
django - rest - swagger == 2.2.0
Django == 2.2
djangorestframework == 3.9.2
feedparser == 5.2.1
flake8 == 3.7.7
freezegun == 0.3.11
gspread == 3.1.0
html5lib == 1.0.1
markdown == 3.1
oauth2client == 4.1.3
Pillow == 6.0.0
pyfcm == 1.4.7
pylint - django == 2.0.6
pylint - runner == 0.5.4
pylint == 2.3.1
psycopg2 - binary == 2.8.2
python - dateutil == 2.8.0
pywebpush == 1.9.3
requests == 2.21.0
pyyaml == 5.1
